chore: split gensql for better readability; renamed create_entries
