--- test.json
@@ -1,27 +1,0 @@
-
-{
-    "user_id": {
-        "type": "bigint unsigned",
-        "nullable": "false",
-        "auto_increment": "true",
-        "primary_key": "true"
-    },
-    "email": {
-        "type": "varchar",
-        "width": "255",
-        "nullable": "true"
-    },
-    "foo_dec": {
-        "type": "decimal",
-        "nullable": "false"
-    },
-    "bar_dbl": {
-        "type": "double",
-        "nullable": "true"
-    },
-    "last_modified": {
-        "type": "timestamp",
-        "nullable": "true",
-        "default": "null"
-    }
-}
